// src/shared/types/db.ts
export namespace DB {
// ...
	export type OngekiProfileData = {
		id: number
		user: number
		version: number
		userName: string | null
		level: number | null
		reincarnationNum: number | null
		exp: number | null
		point: number | null
		totalPoint: number | null
		playCount: number | null
		jewelCount: number | null
		totalJewelCount: number | null
		medalCount: number | null
		playerRating: number | null
		highestRating: number | null
		battlePoint: number | null
		nameplateId: number | null
		trophyId: number | null
		cardId: number | null
		characterId: number | null
		characterVoiceNo: number | null
		tabSetting: number | null
		tabSortSetting: number | null
		cardCategorySetting: number | null
		cardSortSetting: number | null
		playedTutorialBit: number | null
		firstTutorialCancelNum: number | null
		sumTechHighScore: number | null
		sumTechBasicHighScore: number | null
		sumTechAdvancedHighScore: number | null
		sumTechExpertHighScore: number | null
		sumTechMasterHighScore: number | null
		sumTechLunaticHighScore: number | null
		sumBattleHighScore: number | null
		sumBattleBasicHighScore: number | null
		sumBattleAdvancedHighScore: number | null
		sumBattleExpertHighScore: number | null
		sumBattleMasterHighScore: number | null
		sumBattleLunaticHighScore: number | null
		eventWatchedDate: string | null
		cmEventWatchedDate: string | null
		firstGameId: string | null
		firstRomVersion: string | null
		firstDataVersion: string | null
		firstPlayDate: string | null
		lastGameId: string | null
		lastRomVersion: string | null
		lastDataVersion: string | null
		compatibleCmVersion: string | null
		lastPlayDate: string | null
		lastPlaceId: number | null
		lastPlaceName: string | null
		lastRegionId: number | null
		lastRegionName: string | null
		lastAllNetId: number | null
		lastClientId: string | null
		lastUsedDeckId: number | null
		lastPlayMusicLevel: number | null
		banStatus: number
		rivalScoreCategorySetting: number
		overDamageBattlePoint: number
		bestBattlePoint: number
		lastEmoneyBrand: number
		lastEmoneyCredit: number
		isDialogWatchedSuggestMemory: boolean
		newHighestRating: number | null
		newPlayerRating: number | null
	}
// ...
	export type OngekiProfileRating = {
		id: number
		user: number
		version: number
		type: string
		index: number
		musicId: number | null
		platinumScoreMax: number
		platinumScoreStar: number
		difficultId: number | null
		romVersionCode: number | null
		techScoreMax: number
	}
// ...
}